time_t is now long long (will certainly require a types compatibility design)

+++ include/sys/types.h
@@ -62,7 +62,7 @@ typedef int ssize_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long long time_t;
 # endif
 
 # ifndef gid_t
